Create common EditFrequentTransactionDesecriptor
